device/transport/mqtt/requirements.txt
# SB-Utils@git+https://github.com/ScreamBun/SB_utils.git#subdirectory=root
paho-mqtt==1.5.1
kombu==5.0.2
python-etcd  # Use version from SB-Utils